Add category search filter
This builds on the exsiting search functionality by adding the ability
of a user to search based on pages and articles, but also by a specific
category. They can also use keywords within that category. Design has
also slightly changed, with the header having a higher height to
accomidate the new search fields.